# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"BigNumber": {},
...
			"User": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					},
					"disabled": {
						"type": "boolean"
					},
					"bio": {
						"type": "string"
					},
					"xpPoints": {
						"type": "number",
						"format": "double"
					},
					"rewardReceiverAddress": {
						"type": "string"
					},
					"wallets": {
						"items": {
							"properties": {
								"createdAt": {
									"type": "number",
									"format": "double"
								},
								"linked": {
									"type": "boolean"
								},
								"nonce": {
									"type": "string"
								},
								"address": {
									"type": "string"
								}
							},
							"required": [
								"createdAt",
								"linked",
								"nonce",
								"address"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"uid",
					"email",
					"displayName",
					"picture",
					"disabled",
					"bio",
					"xpPoints",
					"rewardReceiverAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersTraitType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layers": {
						"items": {
							"$ref": "#/components/schemas/TraitLayer"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layers"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TraitLayer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layer": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MastersTraitType"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TraitGender": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersColor": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"color": {
						"type": "string"
					},
					"typeId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"color",
					"typeId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersTrait": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.926Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.926Z"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"frontImage": {
						"type": "string"
					},
					"backImage": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MastersTraitType"
					},
					"gender": {
						"$ref": "#/components/schemas/TraitGender"
					},
					"color": {
						"$ref": "#/components/schemas/MastersColor"
					},
					"hide": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"type",
					"gender",
					"color",
					"hide"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersMintTrait": {
				"properties": {
					"colors": {
						"items": {
							"$ref": "#/components/schemas/MastersColor"
						},
						"type": "array"
					},
					"traits": {
						"items": {
							"$ref": "#/components/schemas/MastersTrait"
						},
						"type": "array"
					},
					"type": {
						"$ref": "#/components/schemas/MastersTraitType"
					}
				},
				"required": [
					"colors",
					"traits",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"layers": {
						"items": {
							"$ref": "#/components/schemas/ItemLayer"
						},
						"type": "array"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/ItemCategory"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"layers",
					"categories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemLayer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"layer": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ItemType"
					}
				},
				"required": [
					"id",
					"name",
					"order",
					"layer",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemCategory": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ItemType"
					}
				},
				"required": [
					"id",
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.937Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"default": "2024-12-06T04:30:34.937Z"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/ItemCategory"
					},
					"gender": {
						"$ref": "#/components/schemas/TraitGender"
					},
					"description": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"frontImage": {
						"type": "string"
					},
					"backImage": {
						"type": "string"
					},
					"maskImage": {
						"type": "string"
					},
					"rarity": {
						"type": "string"
					},
					"effects": {
						"type": "string"
					},
					"storyNotes": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"category",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersLootboxJoepegsSignature": {
				"properties": {
					"receiver": {
						"type": "string"
					},
					"tokenIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"amounts": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/MastersItem"
						},
						"type": "array"
					},
					"nonce": {
						"properties": {
							"type": {
								"type": "string"
							},
							"hex": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"hex"
						],
						"type": "object"
					},
					"signature": {
						"type": "string"
					}
				},
				"required": [
					"receiver",
					"tokenIds",
					"amounts",
					"items",
					"nonce",
					"signature"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersLootbox": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"image": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"gameId": {
						"type": "string"
					},
					"genderId": {
						"type": "number",
						"format": "double"
					},
					"ticketId": {
						"type": "number",
						"format": "double"
					},
					"itemWeights": {
						"items": {
							"$ref": "#/components/schemas/MastersLootboxItem"
						},
						"type": "array"
					},
					"prices": {
						"items": {
							"$ref": "#/components/schemas/LootboxPrice"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"active",
					"image",
					"order",
					"chainId",
					"genderId",
					"ticketId",
					"itemWeights",
					"prices"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersLootboxItem": {
				"properties": {
					"weight": {
						"type": "number",
						"format": "double"
					},
					"lootbox": {
						"$ref": "#/components/schemas/MastersLootbox"
					},
					"item": {
						"$ref": "#/components/schemas/MastersItem"
					}
				},
				"required": [
					"weight",
					"lootbox",
					"item"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LootboxPrice": {
				"properties": {
					"lootbox": {
						"$ref": "#/components/schemas/MastersLootbox"
					},
					"currency": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"currency",
					"price",
					"decimals",
					"address",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MastersLootboxSignature": {
				"properties": {
					"receiver": {
						"type": "string"
					},
					"tokenIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"amounts": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/MastersItem"
						},
						"type": "array"
					},
					"nonce": {
						"properties": {
							"type": {
								"type": "string"
							},
							"hex": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"hex"
						],
						"type": "object"
					},
					"claimableUntil": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					},
					"payWithTicket": {
						"type": "boolean"
					},
					"ticketId": {
						"type": "number",
						"format": "double"
					},
					"signature": {
						"type": "string"
					}
				},
				"required": [
					"receiver",
					"tokenIds",
					"amounts",
					"items",
					"nonce",
					"claimableUntil",
					"currency",
					"price",
					"decimals",
					"payWithTicket",
					"ticketId",
					"signature"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WalletSignatureMessage": {
				"properties": {
					"message": {
						"type": "string"
					},
					"nonce": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"nonce"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthCustomToken": {
				"properties": {
					"message": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "masters-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/masters/colors": {
			"get": {
				"operationId": "GetMastersColors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersColor"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "typeId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/genders": {
			"get": {
				"operationId": "GetMastersGenders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TraitGender"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/equip/signature": {
			"post": {
				"operationId": "GetMastersEquipSignature",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"signature": {
											"type": "string"
										},
										"nonce": {
											"$ref": "#/components/schemas/BigNumber"
										},
										"tokenId": {
											"type": "number",
											"format": "double"
										},
										"itemIds": {
											"items": {
												"type": "number",
												"format": "double"
											},
											"type": "array"
										},
										"owner": {
											"type": "string"
										}
									},
									"required": [
										"signature",
										"nonce",
										"tokenId",
										"itemIds",
										"owner"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tokenId": {
										"type": "number",
										"format": "double"
									},
									"owner": {
										"type": "string"
									},
									"itemIds": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"tokenId",
									"owner",
									"itemIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/masters/mint/traits": {
			"post": {
				"operationId": "GetMintTraits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersMintTrait"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/masters/lootbox/buy": {
			"post": {
				"operationId": "BuyLootbox",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"saveId": {
										"type": "string"
									},
									"currency": {
										"type": "string"
									},
									"amount": {
										"type": "number",
										"format": "double"
									},
									"lootboxId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"saveId",
									"currency",
									"amount",
									"lootboxId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/masters/lootbox/joepegs": {
			"post": {
				"operationId": "GetMastersLootboxJPSignature",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MastersLootboxJoepegsSignature"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"receiver": {
										"type": "string"
									},
									"amount": {
										"type": "number",
										"format": "double"
									},
									"lootboxId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"receiver",
									"amount",
									"lootboxId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/masters/lootbox": {
			"get": {
				"operationId": "GetMastersLootboxes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MastersLootbox"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "gameId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "GetMastersLootboxSignature",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MastersLootboxSignature"
								}
							}
						}
					}
				},
				"tags": [
					"Masters"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"payWithTicket": {
										"type": "boolean"
									},
									"currency": {
										"type": "string"
									},
									"amount": {
										"type": "number",
										"format": "double"
									},
									"receiver": {
										"type": "string"
									},
									"lootboxId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"currency",
									"amount",
									"receiver",
									"lootboxId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/auth/wallet": {
			"get": {
				"operationId": "GetWalletAuthMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletSignatureMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostWalletAuthSignature",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthCustomToken"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"signature": {
										"type": "string"
									},
									"address": {
										"type": "string"
									}
								},
								"required": [
									"signature",
									"address"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}